Moves metadata type char to a separate component #483
